Moved cameraMap.js code to Home component
